backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
requests==2.31.0
python-multipart==0.0.6
openpyxl==3.1.2
python-dotenv==1.0.0
google-generativeai==0.3.2
beautifulsoup4==4.12.2
lxml==4.9.3
